Move the OEmbed logic to a separate object.
